entity exists logic

+++ application/src/main/java/com/example/service/ItemService.java
@@ -7,7 +7,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -36,6 +35,10 @@ public class ItemService {
     }
 
     public Page<Item> getFavouriteItems(String email, String name, String category, Pageable pageable) {
+        if (!detailsService.isUserExists(email)) {
+            throw new EntityNotFoundException("Пользователя с таким идентификатором не существует");
+        }
+
         long total = itemRepository.countFavouriteItems(email, name, category);
         List<Item> items = itemRepository.findFavouriteItems(
                 email,
@@ -48,20 +51,39 @@ public class ItemService {
     }
 
     public boolean isFavourite(String email, long id) {
+        if (!detailsService.isUserExists(email)) {
+            throw new EntityNotFoundException("Пользователя с таким идентификатором не существует");
+        }
+        if (!this.isItemExists(id)) {
+            throw new EntityNotFoundException("Предмета с таким идентификатором не существует");
+        }
+
         return itemRepository.isFavourite(email, id);
     }
 
     public void addFavouriteItem(String username, long id) {
-        UserDetails user = detailsService.loadUserByUsername(username);
-        Item item = this.getItem(id);
+        if (!detailsService.isUserExists(username)) {
+            throw new EntityNotFoundException("Пользователя с таким идентификатором не существует");
+        }
+        if (!this.isItemExists(id)) {
+            throw new EntityNotFoundException("Предмета с таким идентификатором не существует");
+        }
 
         itemRepository.addFavouriteItem(username, id);
     }
 
     public void deleteFavouriteItem(String username, long id) {
-        UserDetails user = detailsService.loadUserByUsername(username);
-        Item item = this.getItem(id);
+        if (!detailsService.isUserExists(username)) {
+            throw new EntityNotFoundException("Пользователя с таким идентификатором не существует");
+        }
+        if (!this.isItemExists(id)) {
+            throw new EntityNotFoundException("Предмета с таким идентификатором не существует");
+        }
 
         itemRepository.deleteFavouriteItem(username, id);
+    }
+
+    public boolean isItemExists(long id) {
+        return itemRepository.isItemExists(id);
     }
 }
